% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_hbv.R
\name{calibrate_HBV}
\alias{calibrate_HBV}
\title{calibrate_HBV()}
\usage{
calibrate_HBV(weatherlist, streamflow, param_limits, score_function = "MSE")
}
\arguments{
\item{score_function}{}
}
\value{

}
\description{
calibrate_HBV()
}
