% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_hbv.R
\name{distribute_weather}
\alias{distribute_weather}
\title{distribute_weather()}
\usage{
distribute_weather(tempvec, precipvec, datevec, basin_data, latitude)
}
\arguments{
\item{latitude}{}
}
\value{

}
\description{
distribute_weather()
}
