% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_hbv.R
\name{hydrological_hbv}
\alias{hydrological_hbv}
\title{hydrological_hbv()}
\usage{
hydrological_hbv(
  basin,
  tair,
  precip,
  pet,
  param_snow,
  param_soil,
  param_route,
  param_tf,
  init_snow = 0,
  init_soil = 0,
  init_routing = c(0, 0, 0),
  returnDetails = FALSE,
  rout_model = 3
)
}
\arguments{
\item{rout_model}{}
}
\value{

}
\description{
hydrological_hbv()
}
