% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_hbv.R
\name{make_basin_data}
\alias{make_basin_data}
\title{make_basin_data}
\usage{
make_basin_data(areas, refh)
}
\arguments{
\item{refh}{The height of location of the precip/temp gauge. If you use gridded data from e.g. seNorge, use mean elevation in the area.}
}
\value{
A table with catchment information on a suitable format for further use.
}
\description{
make_basin_data
}
