% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_hbv.R
\name{obj_fun_hbv}
\alias{obj_fun_hbv}
\title{obj_fun_hbv()}
\usage{
obj_fun_hbv(
  params,
  weatherlist,
  qq,
  rm = -c(1:365),
  score_function = "MSE",
  nsw = 0.3,
  rout_model = 3
)
}
\arguments{
\item{rout_model}{}
}
\value{

}
\description{
obj_fun_hbv()
}
