% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{prepare_data}
\alias{prepare_data}
\title{prepare_data}
\usage{
prepare_data(dt, maxgap = 5)
}
\arguments{
\item{dt}{A data table with columns date (daily resolution) and value. The value includes data, e.g.  streamflow, precip or temp, that should be imputed/prepared. Other columns than value and date can also be present.}

\item{maxgap}{Maximum day gap allowed for imputation. If the gap is larger than maxgap, missing values are not predicted/imputed. Instead, a data table with the longest data sequence is returned.}
}
\value{
The data table with missing values imputed values. If there are too many missing values in a row, the longest sequence of data without missing values is returned.
}
\description{
Imputes missing values and/or return the longest sequence of non-missing data. Imputation method: Linear interpolation through the imputeTS package.
}
