% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_hbv.R
\name{sim_HBV}
\alias{sim_HBV}
\title{sim_HBV()}
\usage{
sim_HBV(
  mypars,
  weatherlist,
  init_snow = 0,
  init_soil = 0,
  init_routing = c(0, 0, 0)
)
}
\arguments{
\item{init_routing}{}
}
\value{

}
\description{
sim_HBV()
}
